pyproject.toml
[tool.poetry]
name = "be-httm"
version = "0.1.0"
description = ""
authors = ["quan.daominh <quan.daominh@ncc.asia>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
fastapi = "^0.115.0"
uvicorn = "^0.30.0"
sqlalchemy = "^2.0"
pymysql = "^1.1.0"
python-multipart = "^0.0.9"
pydantic = "^2.7.0"
pillow = "^10.3.0"
torch = "^2.8.0"
opencv-python = "^4.10.0"
pyjwt = "^2.8.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
